Botones toolbar y cambio del botón de cerrar sesión
