+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:supportsRtl="true"
         android:theme="@style/MyTheme.DayNight"
         tools:ignore="GoogleAppIndexingWarning">
-        <activity android:name=".AccountActivity" />
+        <activity android:name=".moduleAccount.view.AccountActivity" />
         <activity
             android:name=".moduleLogin.view.MainActivity"
             android:theme="@style/Splash">
